Build project using esbuild
This is much more flexible than `monaco-plugin-helpers`. Because of
this, it’s possible to use Prettier as an external dependency, so users
can ignore it in their own build process of they choose not to support
formatting.

Instead of rewriting YAML language service imports in the build tool, I
decided to just import them from where they need to be imported in the
source code.

Closes #63